backend/requirements.txt
# ── Core ──────────────────────────────────────────────
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework-simplejwt==5.3.1

# ── Database ──────────────────────────────────────────
# psycopg2-binary==2.9.9
django-redis==5.4.0

# ── Celery ────────────────────────────────────────────
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
redis==5.0.6

# ── Crawler ───────────────────────────────────────────
httpx==0.27.0
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.2.2

# ── AI ────────────────────────────────────────────────
anthropic==0.28.0

# ── Utils ─────────────────────────────────────────────
Pillow==10.4.0
python-decouple==3.8
gunicorn==22.0.0
